Add multivac and multivac k8s configurations
multivac.yaml contains the deployment configuration for multivac.

multivac-service.yaml contains the service configuration for multivac.
This configuration contains a LoadBalancer service type in order to use
the load balancer of the cloud provider for the incomming requests and
to have an external IP. This
service definition also includes a port redirect from port 80 to port
5000.

multivac-worker.yaml contains the deployment configuration for
multivac-workers.