Add redis kubernetes configuration
redis.yaml contains the deployment configuration for the redis db. This
configuration declares a volume as well as a volume claim. (WARNING:
This configuration is not taking security into account since the cluster
is not password protected).

redis-volumeclaim.yaml contains the volume claim defition for redis in a
ReadWriteOnce mode.

redis-service.yaml contains the service definition for redis.